Extract Operation from core.py
